+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.7.3]
+- shortcut methods for create, update and replace
+- In, AllIn and AnyIn now should work in queries
+
 ## [0.7.2] - Range and Filter
 - we're now using a QueryPlanner that analyses the indexes
 - changed View::default() so it's much leaner and works better